+++ test/dummy/app/controllers/api/v1/everything_controller.rb
@@ -11,25 +11,27 @@ class API::V1::EverythingController < API::V1::BaseController
       test them.
     MD
 
-    param :file1, [:file, max_size: 2.megabytes, content_types: %w[image/jpeg image/png]]
+    param :file1, [ :file, max_size: 2.megabytes, content_types: %w[image/jpeg image/png] ]
 
-    param :string1, [:string, empty: false, strip: true, min_length: 1, max_length: 100], default: "Foo"
+    param :string1, [ :string, empty: false, strip: true, min_length: 1, max_length: 100 ], default: "Foo"
 
-    param :integer1, [:integer, min: 1, max: 100], default: 5
+    param :integer1, [ :integer, min: 1, max: 100 ], default: 5
+
+    param :float1, [ :float, min: 0, max: 50 ], default: 1.2
 
     param :hash1, [
       :hash,
-      [:string, empty: false],
-      [:description, "A description of the hash values", [:array, [:integer, min: 0, max: 10]]]
+      [ :string, empty: false ],
+      [ :description, "A description of the hash values", [ :array, [ :integer, min: 0, max: 10 ] ] ]
     ]
 
     param :agreement1, :agreement
 
-    param :big_decimal1, [:big_decimal, min: 0, max: 100]
+    param :big_decimal1, [ :big_decimal, min: 0, max: 100 ]
 
     param :boolean1, :boolean
 
-    param :date1, [:date, min: -> { 1.week.ago }, max: -> { Date.today }]
+    param :date1, [ :date, min: -> { 1.week.ago }, max: -> { Date.today } ]
 
     param :date_range1, [
       :date_range,
@@ -51,11 +53,11 @@ class API::V1::EverythingController < API::V1::BaseController
 
     param :date_time_unix_epoch, :date_time_unix_epoch
 
-    param :enum1, [:enum, %w[one two three]]
+    param :enum1, [ :enum, %w[one two three] ]
 
     response 200, { message: "ok" }
   end
-  
+
   def create
     validated_data = Request.validate!(params)